Moving eventsim changes to eventsim directory
